+++ lib/src/models/basic_time_selector_style.dart
@@ -1,7 +1,17 @@
 import 'package:basic_selector/src/models/models.dart';
 import 'package:flutter/material.dart';
 
+/// The style configuration for [BasicTimeSelector]
 class BasicTimeSelectorStyle extends BasicSelectorStyle {
+  /// The [TextStyle] for the two dot (`:`) separator that is in the middle of the two selectors
+  ///
+  /// Defaults to:
+  /// ```dart
+  /// const TextStyle(
+  ///   fontWeight: FontWeight.bold,
+  ///   fontSize: 20,
+  /// )
+  /// ```
   final TextStyle twoDotSeparatorStyle;
 
   const BasicTimeSelectorStyle({