+++ CHANGELOG.md
@@ -13,3 +13,11 @@
 ## 0.0.2
 
 * Fix constructor was missing super parameters
+
+## 0.1.0
+
+* Added new parameters for `BasicTimeSelectorDialogConfig`: 
+    - `padding` Padding for the actions and the content
+    - `builder` To build a custom dialog with custom content
+    - `extraContent` Extra content below the selector
+* Added new examples to example app